# Use local pcache with full type name as key