--- a/go.mod
+++ b/go.mod
@@ -2,9 +2,6 @@ module github.com/goodbadreviewer/immcheck
 
 go 1.18
 
-require (
-	github.com/storozhukBM/pcache v1.0.1
-	github.com/zeebo/xxh3 v1.0.2
-)
+require github.com/zeebo/xxh3 v1.0.2
 
 require github.com/klauspost/cpuid/v2 v2.0.9 // indirect