Use local pcache with full type name as key

No perf degradation
```
name                                          old time/op    new time/op    delta
ImmcheckBytes/[16384]byte;muts(0%)-8            4.21µs ±33%    4.03µs ±29%   ~     (p=1.000 n=5+5)
ImmcheckBytes/[16384]byte;muts(1%)-8            3.36µs ±27%    2.92µs ±19%   ~     (p=0.222 n=5+5)
ImmcheckTransactions/[8]txs(8);muts(0%)-8       83.8µs ± 1%    84.0µs ± 1%   ~     (p=0.548 n=5+5)
ImmcheckTransactions/[8]txs(8);muts(1%)-8       84.0µs ± 1%    84.1µs ± 1%   ~     (p=0.548 n=5+5)
ImmcheckTransactions/[8]txs(1024);muts(0%)-8    5.92ms ± 1%    5.90ms ± 2%   ~     (p=0.841 n=5+5)
ImmcheckTransactions/[8]txs(1024);muts(1%)-8    5.90ms ± 1%    5.92ms ± 1%   ~     (p=0.421 n=5+5)

name                                          old muts       new muts       delta
ImmcheckBytes/[16384]byte;muts(0%)-8              0.00           0.00        ~     (all equal)
ImmcheckBytes/[16384]byte;muts(1%)-8             4.75k ± 3%     4.77k ± 2%   ~     (p=0.841 n=5+5)
ImmcheckTransactions/[8]txs(8);muts(0%)-8         0.00           0.00        ~     (all equal)
ImmcheckTransactions/[8]txs(8);muts(1%)-8          140 ±12%       151 ±20%   ~     (p=0.452 n=5+5)
ImmcheckTransactions/[8]txs(1024);muts(0%)-8      0.00           0.00        ~     (all equal)
ImmcheckTransactions/[8]txs(1024);muts(1%)-8     1.60 ±100%      2.00 ± 0%   ~     (p=0.794 n=5+4)

name                                          old alloc/op   new alloc/op   delta
ImmcheckBytes/[16384]byte;muts(0%)-8             0.00B          0.00B        ~     (all equal)
ImmcheckBytes/[16384]byte;muts(1%)-8             0.00B          0.00B        ~     (all equal)
ImmcheckTransactions/[8]txs(8);muts(0%)-8         563B ± 1%      563B ± 3%   ~     (p=0.317 n=5+5)
ImmcheckTransactions/[8]txs(8);muts(1%)-8         564B ± 2%      566B ± 1%   ~     (p=0.413 n=5+5)
ImmcheckTransactions/[8]txs(1024);muts(0%)-8    5.91kB ± 1%    5.90kB ± 2%   ~     (p=0.841 n=5+5)
ImmcheckTransactions/[8]txs(1024);muts(1%)-8    6.00kB ± 2%    5.93kB ± 2%   ~     (p=0.421 n=5+5)

name                                          old allocs/op  new allocs/op  delta
ImmcheckBytes/[16384]byte;muts(0%)-8              0.00           0.00        ~     (all equal)
ImmcheckBytes/[16384]byte;muts(1%)-8              0.00           0.00        ~     (all equal)
ImmcheckTransactions/[8]txs(8);muts(0%)-8         3.00 ± 0%      3.00 ± 0%   ~     (all equal)
ImmcheckTransactions/[8]txs(8);muts(1%)-8         3.00 ± 0%      3.00 ± 0%   ~     (all equal)
ImmcheckTransactions/[8]txs(1024);muts(0%)-8      20.0 ± 0%      20.0 ± 0%   ~     (all equal)
ImmcheckTransactions/[8]txs(1024);muts(1%)-8      21.0 ± 0%      20.4 ± 3%   ~     (p=0.095 n=4+5)
```